check file exists in failed bucket before deleting it from the processing bucket

+++ crossword-xml-uploader/src/main/scala/com/gu/crossword/crosswords/CrosswordStore.scala
@@ -43,6 +43,8 @@ trait CrosswordStore {
   def archiveFailedCrosswordXMLFile(config: Config, awsKey: String): Unit = {
     val processingFailedBucketName = "crossword-failed-files"
     s3Client.copyObject(config.crosswordsBucketName, awsKey, processingFailedBucketName, awsKey)
-    s3Client.deleteObject(config.crosswordsBucketName, awsKey)
+    if(s3Client.doesObjectExist(processingFailedBucketName, awsKey)) {
+      s3Client.deleteObject(config.crosswordsBucketName, awsKey)
+    }
   }
 }
